Modify function placing
