+++ torch/csrc/jit/passes/onnx/remove_inplace_ops_for_onnx.h
@@ -9,7 +9,5 @@ TORCH_API void RemoveInplaceOpsForONNX(const std::shared_ptr<Graph>& graph);
 
 TORCH_API void PrepareInplaceOpsForONNX(const std::shared_ptr<Graph>& graph);
 
-TORCH_API void ImplicitCastForBinaryInplaceOps(Block* block);
-
 } // namespace jit
 } // namespace torch
