+++ torch/csrc/jit/passes/remove_inplace_ops.cpp
@@ -1,4 +1,3 @@
-#include <torch/csrc/jit/passes/onnx/remove_inplace_ops_for_onnx.h>
 #include <torch/csrc/jit/passes/remove_inplace_ops.h>
 
 namespace torch {
@@ -76,6 +75,53 @@ void RemoveInplaceOps(Block* block) {
   }
 }
 } // namespace
+
+// Handles special case of binary inplace ops, where the first input node
+// has a lower type precedence than the second input node. When the
+// inplace node is converted to a regular op, this information is lost and
+// the resulting type is based on type precedence, just like regular ops.
+// To avoid this loss of information, we add a cast node before the input
+// node with the higher data type precedence, so that both the input types
+// are the same.
+// An example scenario would be:
+// Before:
+// graph(%0 : Float),
+//        %1 : Half):
+//   %4 : Float = onnx::Cast[to=1](%1)
+//   %5 : Float = onnx::Add(%4, %0)
+//   ...
+// After:
+// graph(%0 : Float),
+//        %1 : Half):
+//   %4 : Half = onnx::Cast[to=10](%0)
+//   %5 : Half = onnx::Add(%1, %4)
+//   ...
+
+void ImplicitCastForBinaryInplaceOps(Block* b) {
+  for (auto it = b->nodes().begin(), end = b->nodes().end(); it != end; ++it) {
+    for (auto* child_block : it->blocks()) {
+      ImplicitCastForBinaryInplaceOps(child_block);
+    }
+
+    // Check type if inplace operation is a binary node
+    if ((it->kind() == aten::add_) || (it->kind() == aten::sub_) ||
+        (it->kind() == aten::mul_) || (it->kind() == aten::div_)) {
+      auto orignalInputs = it->inputs();
+      TensorTypePtr firstInp_tensor =
+          orignalInputs.at(0)->type()->cast<TensorType>();
+      TensorTypePtr secondInp_tensor =
+          orignalInputs.at(1)->type()->cast<TensorType>();
+      if ((firstInp_tensor) && (secondInp_tensor) &&
+          (firstInp_tensor->scalarType().has_value())) {
+        auto newInputNode = it->owningGraph()->create(aten::type_as, 1);
+        newInputNode->insertBefore(*it);
+        newInputNode->addInput(orignalInputs.at(1));
+        newInputNode->addInput(orignalInputs.at(0));
+        it->replaceInput(1, newInputNode->outputs().at(0));
+      }
+    }
+  }
+}
 
 void RemoveInplaceOps(const std::shared_ptr<Graph>& graph) {
   ImplicitCastForBinaryInplaceOps(graph->block());
