+++ torch/csrc/jit/passes/remove_inplace_ops.h
@@ -8,5 +8,7 @@ namespace torch {
 namespace jit {
 // see .cpp for docs
 TORCH_API void RemoveInplaceOps(const std::shared_ptr<Graph>& graph);
+
+TORCH_API void ImplicitCastForBinaryInplaceOps(Block* block);
 } // namespace jit
 } // namespace torch
